Rename graph creation functions
