+++ kglib/utils/graph/create/concept_dict_to_graph_test.py
@@ -16,11 +16,12 @@
 #  specific language governing permissions and limitations
 #  under the License.
 #
+
 import unittest
 import networkx as nx
 from kglib.kgcn.core.ingest.traverse.data.context.neighbour import Thing
 
-from kglib.utils.graph.create.model.standard.convert import concept_dict_to_graph
+from kglib.utils.graph.create.concept_dict_to_graph import concept_dict_to_graph
 from kglib.utils.graph.test.case import GraphTestCase
