--- kglib/utils/graph/create/model/standard/BUILD
@@ -1,26 +1,0 @@
-load("@io_bazel_rules_python//python:python.bzl", "py_test", "py_library")
-load("@pypi_dependencies//:requirements.bzl", "requirement")
-
-
-py_test(
-    name = "convert_test",
-    srcs = [
-        "convert_test.py"
-    ],
-    deps = [
-        "standard",
-        "//kglib/graph/test"
-    ]
-)
-
-py_library(
-    name = "standard",
-    srcs = glob(['**/*.py']),
-    deps = [
-        requirement('networkx'),
-        requirement('decorator'),
-        '//kglib/kgcn',
-        '//kglib/graph/create/model/common'
-    ],
-    visibility=['//visibility:public']
-)
