+++ kglib/utils/graph/create/queries_to_graph_test.py
@@ -21,7 +21,7 @@ import unittest
 
 import networkx as nx
 
-from kglib.utils.graph.create.from_queries import concept_dict_from_concept_map, combine_2_graphs
+from kglib.utils.graph.create.queries_to_graph import concept_dict_from_concept_map, combine_2_graphs
 from kglib.utils.graph.test.mock.answer import MockConceptMap
 from kglib.utils.graph.test.mock.concept import MockType, MockThing
 from kglib.utils.graph.test.case import GraphTestCase
